Add staking pool2 to protocols
